refactor+test: apply code-review fixes
- Extract the duplicated loadUsers into a shared job helper (introduced during
  the /simplify efficiency pass, so the simplify agents didn't see it).
- Add a digest int test pinning the intended spec separation: a registrant's
  reminder-unsubscribe does NOT hide their registration from the manager digest
  (unsubscribe covers reminders only, per #589).

Dismissed review findings (verified against source/spec):
- send-then-persist ordering (reminder ledger / digest watermark) is the same
  accepted at-least-once trade-off as ExpireEvents; persist-then-send would risk
  missing a send. Normal re-runs/retries are dedup-proven by the int specs.
- digest correctly omits _status / remindersUnsubscribedAt filters: it's the
  batched manager notification (unsubscribe is registrant-reminder-only), and
  filtering _status would miss real registrations for later-unpublished events.
- recipient-null watermark stall is unreachable (managers are auth users with a
  required email; pickChannel falls back to it).

Refs #589
